Fixed authentication in front-office

+++ public/front-office/src/app/services/categories.service.ts
@@ -1,3 +1,6 @@
+import { NotificationsService } from './notifications.service';
+import { Router } from '@angular/router';
+import { TokenService } from './token.service';
 import { Injectable } from '@angular/core';
 import { HttpService } from './http.service';
 
@@ -6,7 +9,11 @@ import { HttpService } from './http.service';
 })
 export class CategoriesService {
 
-    constructor(public http: HttpService) { }
+    constructor(public http: HttpService,
+                private tokenService: TokenService,
+                private router: Router,
+                private notificationsService: NotificationsService) { }
+
 
     getCategories() {
         return this.http.get('/categories');
@@ -17,7 +24,13 @@ export class CategoriesService {
     }
 
     deleteCategory(id) {
-        return this.http.delete('/categories/'+id)
+        if (this.tokenService.isTokenSet())  
+            return this.http.delete('/categories/'+id)
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
+
     }
 
     getItems(id) {
