+++ public/front-office/src/app/services/items.service.ts
@@ -1,3 +1,6 @@
+import { Router } from '@angular/router';
+import { NotificationsService } from './notifications.service';
+import { TokenService } from './token.service';
 import { Injectable } from '@angular/core';
 import { HttpService } from './http.service';
 
@@ -6,10 +9,13 @@ import { HttpService } from './http.service';
 })
 export class ItemsService {
 
-    constructor(public http: HttpService) { }
+    constructor(public http: HttpService,
+                private tokenService: TokenService,
+                private notificationsService: NotificationsService,
+                private router: Router) { }
 
     getItems() {
-        return this.http.get('/items');
+        return this.http.get('/items') 
     }
 
     getItemById(id) {
@@ -17,10 +23,32 @@ export class ItemsService {
     }
 
     checkIfAvailable(startDate, endDate, objectId) {
-        return this.http.post('/items/'+objectId+'/checkIfAvailable', {startDate, endDate, objectId})
+        if (this.tokenService.isTokenSet())  
+            return this.http.post('/items/'+objectId+'/checkIfAvailable', {startDate, endDate, objectId})
+            else {
+                this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+                this.router.navigate(['/login']);
+            }
+
+    }
+
+    createItem(item) {
+        if (this.tokenService.isTokenSet())  
+            return this.http.post('/items', item);
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
+
     }
 
     deleteItem(id) {
-        return this.http.delete('/items/'+id)
+        if (this.tokenService.isTokenSet())  
+            return this.http.delete('/items/'+id)
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
+
     }
 }
