+++ public/front-office/src/app/services/kits.service.ts
@@ -1,3 +1,6 @@
+import { Router } from '@angular/router';
+import { NotificationsService } from './notifications.service';
+import { TokenService } from './token.service';
 import { Injectable } from '@angular/core';
 import { HttpService } from './http.service';
 
@@ -6,7 +9,10 @@ import { HttpService } from './http.service';
 })
 export class KitsService {
 
-    constructor(public http: HttpService) { }
+    constructor(public http: HttpService,
+                private tokenService: TokenService,
+                private notificationsService: NotificationsService,
+                private router: Router) { }
 
     getKits() {
         return this.http.get('/Kits');
@@ -17,14 +23,29 @@ export class KitsService {
     }
 
     checkIfAvailable(startDate, endDate,kitId, objectId) {
-        return this.http.post('/Kits/'+kitId+'/checkIfAvailable', {startDate, endDate, objectId})
+        if (this.tokenService.isTokenSet())  
+            return this.http.post('/Kits/'+kitId+'/checkIfAvailable', {startDate, endDate, objectId})
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
     }
 
     deleteKit(id) {
-        return this.http.delete('/Kits/'+id)
+        if (this.tokenService.isTokenSet())  
+            return this.http.delete('/Kits/'+id)
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
     }
 
     createRental(startDate, endDate, timeInDays, objectId, kitId) {
-        return this.http.post('/rentals', {startDate, endDate, timeInDays, objectId, rentalType:"prenotazione",rentalTarget: 'kit',state:'futura', kitId});
+        if (this.tokenService.isTokenSet())  
+            return this.http.post('/rentals', {startDate, endDate, timeInDays, objectId, rentalType:"prenotazione",rentalTarget: 'kit',state:'futura', kitId});
+        else {
+            this.notificationsService.error("Devi essere autenticato per effettuare questa operazione.")
+            this.router.navigate(['/login']);
+        }
     }
 }
